@@ -1,6 +1,6 @@
 {
   "name": "@tachybase/build",
-  "version": "1.0.22",
+  "version": "1.2.6",
   "description": "Library build tool based on tsup.",
   "license": "Apache-2.0",
   "main": "lib/index.js",
@@ -15,7 +15,6 @@
   },
   "dependencies": {
     "@hapi/topo": "^6.0.2",
-    "@pnpm/find-workspace-dir": "^6.0.3",
     "@pnpm/workspace.find-packages": "^1.1.12",
     "@rsbuild/plugin-less": "1.1.0",
     "@rsbuild/plugin-react": "^1.1.0",
@@ -29,11 +28,10 @@
     "fs-extra": "^11.2.0",
     "tar": "^6.2.1",
     "tsup": "8.3.5",
-    "typescript": "5.7.2",
+    "typescript": "5.8.3",
     "update-notifier": "3.0.1",
     "vite": "^5.4.11",
     "vite-plugin-css-injected-by-js": "^3.5.2",
-    "vite-plugin-lib-inject-css": "1.3.0",
     "yargs-parser": "13.1.2"
   },
   "devDependencies": {
