@@ -1,6 +1,6 @@
 {
   "name": "@tachybase/plugin-print-template",
-  "version": "1.0.22",
+  "version": "1.2.6",
   "main": "dist/server/index.js",
   "scripts": {
     "build": "tachybase-build --no-dts @tachybase/plugin-print-template"
@@ -14,7 +14,7 @@
     "koa-bodyparser": "^4.4.1",
     "pizzip": "^3.1.7",
     "ts-node": "10.9.2",
-    "typescript": "^5.7.2",
+    "typescript": "5.8.3",
     "xml2js": "^0.6.2"
   },
   "peerDependencies": {
