@@ -1,4 +1,4 @@
-import { i18n } from '@tachybase/client';
+import { i18n, tval as nTval } from '@tachybase/client';
 
 import { useTranslation } from 'react-i18next';
 
@@ -13,7 +13,10 @@ export function generateNTemplate(key: string) {
 }
 
 export function useDuplicatorTranslation() {
-  return useTranslation(NAMESPACE, {
+  return useTranslation([NAMESPACE, 'core'], {
     nsMode: 'fallback',
   });
 }
+
+export const tval = (key: string) => nTval(key, { ns: NAMESPACE });
+export { useTranslation };
