@@ -3,15 +3,17 @@ import { ExtendCollectionsProvider, SchemaComponent, SchemaComponentContext, use
 import { onFieldChange, useField, useFormEffects } from '@tachybase/schema';
 
 import WorkflowPlugin, { RadioWithTooltip } from '.';
+import { AddWorkflowCategory, AddWorkflowCategoryAction } from './components/AddWorkflowCategory';
 import { ColumnShowCollection } from './components/ColumnShowCollection';
 import { ColumnShowTitle } from './components/ColumnShowTitle';
+import { EditWorkflowCategory, EditWorkflowCategoryAction } from './components/EditWorkflowCategory';
 import { ExecutionLink } from './components/ExecutionLink';
 import { ExecutionRetryAction } from './components/ExecutionRetryAction';
 import { ExecutionStatusColumn, ExecutionStatusSelect } from './components/ExecutionStatus';
 import OpenDrawer from './components/OpenDrawer';
 import { ExecutionResourceProvider } from './ExecutionResourceProvider';
 import { executionCollection } from './schemas/executions';
-import { collectionWorkflows, workflowSchema } from './schemas/workflows';
+import { collectionWorkflowCategories, collectionWorkflows, workflowSchema } from './schemas/workflows';
 import { WorkflowLink } from './WorkflowLink';
 
 function SyncOptionSelect(props) {
@@ -41,7 +43,7 @@ export function WorkflowPane(props) {
 
   const { getTriggersOptions } = usePlugin(WorkflowPlugin);
   return (
-    <ExtendCollectionsProvider collections={[collectionWorkflows, executionCollection]}>
+    <ExtendCollectionsProvider collections={[collectionWorkflows, executionCollection, collectionWorkflowCategories]}>
       <SchemaComponentContext.Provider value={{ ...ctx, designable: false }}>
         <SchemaComponent
           schema={schema}
@@ -55,6 +57,10 @@ export function WorkflowPane(props) {
             ExecutionStatusColumn,
             ColumnShowTitle,
             ColumnShowCollection,
+            AddWorkflowCategory,
+            AddWorkflowCategoryAction,
+            EditWorkflowCategory,
+            EditWorkflowCategoryAction,
             ...components,
           }}
           scope={{
