@@ -1,3 +1,4 @@
+import path from 'node:path';
 import { requireModule } from '@tachybase/utils';
 
 const arr2obj = (items: any[]) => {
@@ -24,7 +25,10 @@ export const getResource = (packageName: string, lang: string, isPlugin = true) 
   }
   for (const prefix of prefixes) {
     try {
-      const file = `${packageName}/${prefix}/locale/${lang}`;
+      let file = `${packageName}/${prefix}/locale/${lang}`;
+      if (isPlugin && process.env.RUN_MODE === 'engine') {
+        file = path.join(process.env.NODE_MODULES_PATH, file);
+      }
       const f = require.resolve(file);
       if (process.env.APP_ENV !== 'production') {
         delete require.cache[f];
