@@ -1,9 +1,18 @@
-import path from 'path';
-import { pathToFileURL } from 'url';
+import path from 'node:path';
+import { pathToFileURL } from 'node:url';
 
 export function requireModule(m: any) {
   if (typeof m === 'string') {
-    m = require(m);
+    if (process.env.RUN_MODE === 'engine') {
+      if (path.isAbsolute(m)) {
+        m = require(m);
+      } else {
+        // adapter to vercel apk environment
+        m = require(path.join(process.env.NODE_MODULES_PATH, m));
+      }
+    } else {
+      m = require(m);
+    }
   }
 
   if (typeof m !== 'object') {
